requirements.txt
asgiref==3.5.2
astroid==2.11.7
beautifulsoup4==4.11.1
Brotli==1.0.9
dill==0.3.5.1
dj-database-url==1.0.0
Django==4.0.4
django-bootstrap4==22.2
django-ckeditor==6.4.2
django-js-asset==2.0.0
gunicorn==20.1.0
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
mysql-connector==2.2.9
mysql-connector-python==8.0.30
mysqlclient==2.1.1
Pillow==9.2.0
platformdirs==2.5.2
protobuf==3.20.1
psycopg2-binary==2.9.3
pylint==2.14.5
pylint-django==2.5.3
pylint-plugin-utils==0.7
PyMySQL==1.0.2
python-decouple==3.6
python-dotenv==0.20.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.11.1
whitenoise==6.2.0
wrapt==1.14.1
